backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
crewai
cerebras_cloud_sdk
pydantic-settings
httpx
